- Fix cast warning, and incorrect unit test checks in ssfstr module.

+++ _codec/ssfstr.c
@@ -282,7 +282,7 @@ tokDone:
     if ((size_t)((cs - (*cstr)) + 1) <= tokenStrSize)
     {
         /* Yes, copy into token buffer */
-        *tokenStrLen = cs - (*cstr);
+        *tokenStrLen = (int32_t) (cs - (*cstr));
         memcpy(tokenStrOut, *cstr, *tokenStrLen);
         tokenStrOut[*tokenStrLen] = 0;
     }
